PIA-831: Refactor and reorganize UI test class structure
I've tried to decouple UiAutomator code from the tests. Which should make it easier if we want to
change the UI testing framework.
The folder structure should now be easier to navigate, where folder objects only contain UiAutomator
objects and its helper methods.
The 'steps' folder contains the abstracted class that manages the sign in steps as well as the specific
UiAutomator steps. If we decide to use a new framework we could extend from SignInSteps and implement
its methods with the new framework.
The tests folder is where the entry point for tests is located (UiSignInTests). If we change the
framework, we should be able to just change the class 'UiSignInTests' extends from and make no further
change to the tests.
'UiAction' contains a further abstraction of the tests. It can be injected with the behavior of the
Ui test steps.