PIA-831: Add step to UI login tests that navigates to the signup screen
Tests should now be able to run in a more consistent way as they all begin from the same screen.
However, when run as a batch, if the test that successfully logs in is executed before the end it will
cause all subsequent tests to fail due to 'navigateToSignUpScreen' not being able to navigate out of the
connection screen. This will be fixed in a following PR. Most likely when implementing logout UI tests