PIA-912: Refactor UI logout tests
This includes loging out of the app from the main screen.
It also includes checking that the persistence is wiped after logout. Which
is failing at the time of this commit.
This commit also includes a refactor of the file structure.
Step interfaces are now under the 'interfaces' folder.
UiAction has been split into different classes to avoid having a class
with multiple responsibilities