PIA-1269: Update VPN Ip field when connecting
In ConnectionViewModel, all calls to 'ConnectionUseCase.startConnection' are now
funneled through the ConnectionViewModel.connect function. So that setIps()
gets called when:
1) Pressing the connect button
2) Pressing quick connect buttons
3) Choosing a region from the regions list
setIps() is now also called when the view is generated. This is to make sure that
even after reconnecting from the settings menu, we show the most recent IP fetched