PIA-1757: Fix bugs when loading splash screen with no internet connection
Bug 1) The splash screen loads the regions list. If there's no connection, then the list
is empty and crashes when attempting to get the best latency server. Now the splash screen
reloads the default list if it is empty.

Bug 2) The splash screen loads the subscription data. If there's no connection, then the
app crashes when connection is resumed and the app tries to access the subscriptions.
Now we load the subscription list when reconnecting to the network if it was empty.
This also reloads the screen and the prices are shown.